* 'develop' of https://github.com/nuls-io/nuls_2.0:
  websocket
  websocket
  websocket
  add transaction model
  consensus and tools commit
  change chainId type
  websocket
  add chain model data (&tx).
  websocket
  websocket

+++ chain/nuls-chain/src/main/java/io/nuls/chain/storage/ChainAssetStorage.java
@@ -1,6 +1,7 @@
 package io.nuls.chain.storage;
 
-import io.nuls.base.data.chain.ChainAsset;
+
+import io.nuls.chain.model.dto.ChainAsset;
 
 import java.util.List;
 
@@ -35,5 +36,5 @@ public interface ChainAssetStorage {
      * @param chainId The chain ID
      * @return ChainAsset object
      */
-    List<ChainAsset> getByChain(short chainId);
+    List<ChainAsset> getByChain(int chainId);
 }
